web/requirements.txt
bcrypt==3.1.4
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.1
Flask-WTF==0.14.2
sqlalchemy-migrate==0.11.0
